@@ -13,25 +13,24 @@
     "ci": "npm run lint && npm run test:once"
   },
   "dependencies": {
-    "@angular/animations": "13.3.11",
-    "@angular/common": "13.3.11",
-    "@angular/compiler": "13.3.11",
-    "@angular/core": "13.3.11",
-    "@angular/forms": "13.3.11",
-    "@angular/platform-browser": "13.3.11",
-    "@angular/platform-browser-dynamic": "13.3.11",
-    "@angular/router": "13.3.11",
+    "@angular/animations": "14.2.5",
+    "@angular/common": "14.2.5",
+    "@angular/compiler": "14.2.5",
+    "@angular/core": "14.2.5",
+    "@angular/forms": "14.2.5",
+    "@angular/platform-browser": "14.2.5",
+    "@angular/platform-browser-dynamic": "14.2.5",
+    "@angular/router": "14.2.5",
     "core-js": "2.6.3",
-    "rxjs": "6.6.7",
+    "rxjs": "7.5.7",
     "tslib": "^2.0.0",
-    "zone.js": "~0.11.4",
-    "rxjs-compat": "6.5.2"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.9",
-    "@angular/cli": "13.3.9",
-    "@angular/compiler-cli": "13.3.11",
-    "@angular/language-service": "13.3.11",
+    "@angular-devkit/build-angular": "^14.2.5",
+    "@angular/cli": "14.2.5",
+    "@angular/compiler-cli": "14.2.5",
+    "@angular/language-service": "14.2.5",
     "@types/chrome": "0.0.197",
     "@types/jasmine": "~3.7.0",
     "@types/node": "^12.11.1",
